pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clearmetrics"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "pyjwt>=2.8.0",
    "pandas>=2.2.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
    "types-PyJWT>=1.7.1",
    "pandas-stubs>=2.2.0",
    "sphinx>=7.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "furo>=2024.1.29",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = false
disallow_untyped_defs = true
disallow_any_generics = true
warn_return_any = true
warn_unused_ignores = true

[tool.setuptools.packages.find]
where = ["."]
include = ["clearmetrics*"]

[tool.coverage.run]
source = ["clearmetrics"]
omit = ["clearmetrics/cli/*", "clearmetrics/ports/*"]

[tool.coverage.report]
fail_under = 90
skip_empty = true
